Removed unused assets and rebuilt asset bundles

+++ GOIMP_Content/Assets/Shaders/AlphaMask.shader
@@ -44,10 +44,10 @@
 			sampler2D _MainTex;
 			sampler2D _MaskTex;
 
-			fixed4 frag (v2f i) : SV_Target
+			float4 frag (v2f i) : SV_Target
 			{
-				fixed4 col = tex2D(_MainTex, i.uv);
-				fixed4 maskCol = tex2D(_MaskTex, i.uv);
+				float4 col = tex2D(_MainTex, i.uv);
+				float4 maskCol = tex2D(_MaskTex, i.uv);
 
 				// Set alpha to mask r value
 				col.a = maskCol.r;
